WIP: Refactoring movetype and terrain_costs (#2486)
WIP: This still feels too complex, but also feels like time to open a PR asking
if anyone wants to treat it as a challenge to improve it. Cherry-picking to the
master branch has non-trivial merge conflicts, list at the bottom of this
commit message.

Simplify the public interface to only the movetype class and a read-only interface
called "terrain_costs", which replaces an older complex class with the same
name. Changes to terrain_costs now need to be done through the movetype
instance, so logic about how movement and vision interact can be handled in
movetype.

The terrain_info class is now a private class implementing the terrain_costs
interface, previously it was a public class with raw pointers to other
instances of itself, and a copy constructor which copied the raw-pointers
as-is. One of the raw pointers is still there, but it's only non-null when both
instances involved are owned by one instance of movetype.

This was started while searching for the cause of issue 2571, although that bug
turned out to be elsewhere. It's now a change that turns assumptions that were
hidden in the classes in to documented scenarios, and hopefully guards them
better.

---- Merge conflicts with master ----

* units.cpp: needs a manual merge. Add the `set_attr_changed(UA_MOVEMENT_TYPE)`
  call to my changes, but in this version it only needs to be added once. My
  change merged five `else if`s in to a single `else if`.
* movetype.cpp: take my version, in each case the conflict is the change of
  std::shared_ptr::unique() to std::shared_ptr::count(), which has been replaced
  with make_data_shareable() and make_data_writable().
* movetype.hpp: take my version, the conflict is a with a minimal change to
  remove the Wdeprecated-copy warning, this commit is an in-depth fix.

For both movetype.cpp and movetype.hpp, the result will be exactly the same as the
1.14 versions, as of:
* origin/1.14 0e38e4217900a07a00ca8b239ae4f06a9164b342
* origin/master 4ed39d7ee7a930636f101bf2175c2e116a312725