Fix a crash playing a replay with an add-on theme that doesn't have button-nextmove
The functionality with this fix still isn't good (the buttons get disabled),
but it's better than crashing, and this one-liner would certainly be included
in any more major fix of the functionality.

A replay that reproduces this crash is linked below. SXRPG 5.2.3 has two replay
themes, of which only one has the button-nextmove.
* https://replays.wesnoth.org/1.14/20190701/SXRPG_TempleOfBones_Turn_8_(53897).bz2