WIP, breaks menu button: replay_controller doesn't need to duplicate logic from the hotkey system
It seems all of the button::enable() calls are unnecessary logic with a
necessary side-effect, because the hotkey system will enable or disable these
buttons based on querying replay_controller::can_execute_command(). However,
the enable() calls had the side-effect of triggering the hotkey system to
check the buttons' required states and redrawing.

WIP: the "Menu" and "Actions" menus can be unresponsive when the replay is paused,
possibly because this is updating the buttons based on their states at a
different time. This might be to do with this snippet, which I removed without
knowing where to replace it.

-	if(stop_condition_->should_stop()) {
-		//user interrupted
-		stop_button()->release();
-	}

The code in handle_generic_event looks dodgy to me, calling add_replay_theme()
for every event except "completely_redrawn". The change is just removing code
that would have been a no-op in the old code, but I don't understand the logic
there.