WIP: wondering if union_type could be removed
WIP: The next paragraph is wrong, it's also used for the minimap.

It seems that all of the union_type / underlying_union_type support is either
to show that type in the help, or for game_board.cpp to add terrains to the
list of encountered terrains. Removing it would remove the reason for having
it in the help, and the code in game_board.cpp could just check both the
movement and defense terrains.