WIP: Make map_data= search for a maps directory in the [binary_path] (#4633)
WIP: Only contains the AToTB, HttT and Tutorial changes so far

Allows both of these to have the same effect:

* map_file=campaigns/Heir_To_The_Throne/maps/01_The_Elves_Besieged.map
* map_file=01_The_Elves_Besieged.map

This allows a lot of copies of the campaign/add-on's name to be omitted. It's
no longer necessary to change every scenario's .cfg file to rename an add-on,
or to move a campaign between UMC and mainline.

This makes [binary_path] a misnomer, as it now also handles a text-based type
of file, however that's going to be the correct path for campaigns or add-ons
that use the standard layout with images/, maps/, scenarios/, etc.