WIP: Refactoring the de-facto saved_game singleton
WIP: It now all seems to work, leaving the comment below until I've tested more,
as it's a checklist of stuff to check.

This can start campaigns, but it can't load existing saves (that jumps to
the end of the campaign instead). Multiplayer untested.

Tests all fail with "Couldn't find [test] with id=" followed by no id.

Refactor savegame::loadgame::set_gamestate to be a private function.
It seemed unnecessary to have this as a separate function, but it might
still be necessary - the config might be affected by WML preprocessor
definitions.