Update the German translation for UtBS
I'm not a fluent speaker, so it's probably inaccurate and I'm not going to
attempt the any of flavor-text in the unit descriptions. But this should
cover all the campaign-specific game mechanics and a few events.

Some of the text was marked fuzzy because of spelling corrections in the
English text, so I've marked those as valid again.