WIP: Refactor the internal state of the CFG name generator
This is under discussion in #5148, but this is an addition to
the discussion rather than a ready-to-merge thing. The bug was that
the `mutable last_` member was making names change based on data that
wasn't controlled by the engine's central RNG. Having refactored a bit
that issue is fixed in a way that makes names in replays be predictable.

The handling of internal_state's `seed` doesn't make sense to me, it looks
like a 640-bit PRNG seed, but it's really just caching 20 calls to the
central RNG and then accessing them as a circular buffer of 20 numbers.