NOT FOR MAINSTREAM: Try and do a two-scenario unit test
This is trying to set up a persistent side's recalls so that [side][unit]
can be tested. It fails differently depending on whether it's run interactively
or not:

When run with --unit, it passes (only part_one runs).

When run with --test, the user has to click "end scenario", the second scenario
then runs with no units on the recall list.

@CelticMinstrel: this should auto-link to #5290