Make the preferences window big enough for HDPI fonts
The hardcoded size of the window was too small for the large fonts on HDPI
displays, this changes to almost-automatic sizing. It leaves an upper limit for
the height of 80% of the main window size, so that there's a border at top and
bottom.

Buttons that were placed at the bottom of pages move up so that they aren't
too far from the other controls. The sub-tab-selector buttons on the
multiplayer tab move above the sub-tabs themselves.

With HDPI text, the checkboxes are placed a little too high to align with
the font, but that's a cosmetic detail; this commit fixes the usability issue.