WIP: Add an "engaged" orb, for units that can't move but can attack (should autolink to #5155 #5523)
WIP: Just a proof of concept to see how it looks.

WIP: The new orb has yellow on the left (the magenta of the .png), and the
disengaged orb is changed so that red is on the left; the images in the in-game
help need to be updated to match. The reason for this is that the engaged orb
will be seen much more often than the disengaged orb, and with the swapped
colors it doesn't as good.

WIP: This includes some clean-up of the disengaged orb implementation,
will cherry-pick those parts from it if the engaged-orb isn't added.

WIP: It would need to be described in Tutorial S2's ZoC explanation. It doesn't
occur in Tutorial S1 as the quintains are level zero.