DEBUGGING: Seeing if the sizes that Pango measures are large enough for Pango
Logged in Wesnoth's tracker as #5785.

In pango_text::calculate_size, once Pango has calculated the size of the text,
take the number that Pango calculated, set that as the maximum width, and ask
Pango if the text fits. Print an error if it makes the text wrap to an extra
line.

This is debugging the GUI2 bug that some text fields are displayed with the
final line missing - and it seems at least one cause of that is a bug in Pango
itself. The sizes that Pango returns are not always large enough to contain the
text.

This commit also has some extra debugging for two specific strings in the
filter-by-tag PR.