DEBUGGING: Log when showing tooltips
Part of debugging #6096, sharing in response to #6304's
https://github.com/wesnoth/wesnoth/pull/6304#issuecomment-972820888

The reason for not adding a warning about overlapping tooltips is that, even
filtering out ones that have the same top-left corner, this logs lots of false
positives. Filtering out matching right-hand corners for right-justified text
didn't solve the problem either (and I dropped that code as it wasn't helping
my debugging).

In display::refresh_report, first the area for the report is cleared (on the
line after the new "Removing tooltips that overlap with" log), and then a loop
that calculates offsets of `x` can add multiple areas from a single report. For
example, "cures, heals+8" has two areas with two tooltips, yet even for the
"cures" part the left edge is a few pixels to the right of the left edge of
`rect` when that log is printed.

The code in `clear_tooltips(const SDLRect&)` doesn't know which tip is which
(although it does have access to the text of the old tip, via `i`, and the
caller has the name of the new tip). When debugging 6096 I just kept an
image-editor open with a screenshot and used it to look up the coordinates.