MAYBE FOR MAINLINE: Put ng namespace on to some string_enums
Game initialisation already has the `ng` namespace. That seems to have
originally been a shorthand for "engine", which isn't descriptive. However,
it's currently exclusively used for declarations in src/game_initialization, so
keeping it as `ng` but saying that's an acronym of "New Game" would fit quite
well.

Many unrelated .cpp files use `#define LOG_NG LOG_STREAM(info, log_engine)`,
where `NG` seems to be more "engine, or miscellanous other stuff". That makes me
feel that it would be wrong to say that `namespace ng` is an acronym of New Game.