BROKEN TEST: add a test for cumulative_leadership
Broken because I tried to implement this as an Alice and Bob test, rather
than using the four-sided `COMMON_KEEP_A_B_C_D_UNIT_TEST`. It's not valid
to give a single unit multiple `[leadership]` abilities that use the same
`id`.

When using the LuaAPI as done here, only the first `[leadership]` is added.
When changing Newfrenchy's test in `test_add_or_sub_in_leadership.cfg`, which
is also an Alice and Bob test, to have multiple abilities with the same `id`s,
the result is also failure; the multiple abilities are shown in the `:inspect`
browser, but the combat results aren't the expected amount.

Going to ask in #6466 about the expected results.