NOT FOR MAINSTREAM: Highlight bounds of text areas, don't render the credits
This puts a green rectangle on the area where text will be drawn. To avoid the
credits crash (#6952), it then skips the actual text rendering if the area is
too big.

When pressing the "Credits" button, this draws a green rectangle in the
top-middle of the screen.

The horizontal size of the credits area is correct - the layout of that screen
has three vertical columns, an intentional layout to put them in the central
third of the screen.

The vertical size is only about 1/5 of the screen, and that's a problem. The
size being passed to draw::fill should cover the whole height of the screen
(after the graphics layer translates it for the drawing-scale vs
rendering-scale), however it seems that draw::fill interprets it proportional
to the height of the scroll area. Even on my landscape monitor, clipped.h >
clipped.w, yet visually it appears wider than it is high. The same affects
draw::blit, so the more text is in the credits the more vertically squashed the
text would be.

This commit distills my findings from trying to reapply f282eb7948's method to
the current canvas.cpp.