WIP: docs and pending review
Planned review
====

`get_specials_and_abilities()` seems to be used in several different ways. That can be solved by refactoring it in a later PR, but adding comments now will help with both reviewing and the later refactor.

If I'm understanding the code correctly, this comment would fit here:
When used for plague abilities, the first ability in the returned list is the one that will be used. TODO more about why this helps.