Repair priority system where only the ability with the highest priority decides the content of the overwriters list
Suppose we have 4 numeric type specials (cth or damage) with the 'value' attribute
a1 has the lowest value and highest priority, but its filters do not match any of the other three
a2 can overwrite a3 but not a4 and a3 can overwrite a4.
a4 having the highest value, it should be used but it is a2 which is displayed because only a1 could overwrite the others with its priority, so a3 remains in the overwriter list and can overwrite a4 in the input list

this commit is here to fix this problem, so that a3 is overwritten in overwriter as it is in input.