wmllint: Stop treating single-hex terrain masks as Wesnoth 1.2 map data
Also change the placement of the newlines in the output. The two changes
are done together so that the changes from running wmllint are together.

If there's only one line, put both double-quotes on the same line:

    mask = "Xv, Xv"

If there's more than one, put the leftmost hex of each row in the same column,
and have a separate line for the closing double-quote:

    mask = "
Xv, Xv
Xv, Xv
"

Prior to Wesnoth 1.3.1, the map data used a single letter per hex instead of
the current comma-separated format. The removed code seems to be for handling
those old maps, and so should only be in wmllint-1.4.

Using `[terrain_mask]mask=Xv` triggered a false identification of the mask as
the old map format, causing it to be converted to `mask=X,v` instead, thus
corrupting data/test/scenarios/behavioral_tests/test_victory_attacks.cfg.