Cherry-pick unit test framework from 1.17.24+dev (7c14bee3f832c16aa1c3ba25b178436963d236e8)
Might as well use the latest, so this is 1.17.24 + the changes for testing
illuminates, hides, skirmisher, teleport and [harm_unit]

Some macros from data/test/scenarios/test_modify_shroud.cfg move to the
framework.

The data/test/macros/pathfind_answers.cfg file isn't included in this backport,
it needs to be 1.16's, because it needs to match 1.16's units.