Fix abilities' [filter_adjacent_location]count= filtering
The Wiki docs say that count= takes a number or numerical range, and
returns true if the number of matches in adjacent locations matches
that range. However, it actually returns false if any location in the
adjacent= set doesn't match, and that seems to have been the behavior
since count= was originally added in 1.13.2.