PROTOTYPE: Detect recursion through filter configs instead of just counting
Instead of counting the number of times that a weapon is participating
in the calculation, this makes copies of the filters and tests whether
a new level of recursion is asking the same question as one that's
already on the stack (for this specific weapon).

This means it has logic similar to newfrenchy83's #9185, combined with
the RAII logic of #9202 for clearing items from the list of open queries.

In terms of performance, for normal weapons, #9185, #9202 (with a
limit of 4) and this are all quite similar.

PROTOTYPE: could a filter ever appear twice on the stack with different
answers, for example if it's checking that the unit is in a backstab
position to two different units?