WIP: Make unit_animation_component::start_animation() support locations
Current status on fake units and animations: move.cpp's
unit_mover::do_move() moves the real unit before getting udisplay.cpp
to animate the move. OTOH, udisplay.cpp's move_unit_between() puts the
fake unit on the starting location, because
unit_animation_component::start_animation() always animates starting
from the unit's current location. This commit is the good part of the
WIP, but along with some experimental stuff that'll be in a different
commit, it's still not animating for later parts of the move.