Format CONTRIBUTING.md's example code with clang-format
Correct the setting of PackConstructorInitializers, as it didn't match
the example; setting it to CurrentLine was worse because my_class'
constructor ended up with everything including the `//` comment on the
same line, so the position no longer made it obvious that the comment
referred to an argument.

This converts spaces to tabs in the example, but it's not just
whitespace changes.

The #include "gettext.hpp" and #include "help/help.hpp" lines are now
sorted alphabetically.

The lambda function is short, thus all on a single line.

Note that we're incrementally moving towards the standard rather than
doing a project-wide reformat that would break the git-blame history.