Test scenario: Give the Shyde a new slows that only works in forest
Weapon specials can have location filters, even if they're applied
directly to the weapon rather than being used like abilities. This
gives the Shyde a special which should be shown as inactive except
when moving to or attacking from a forest hex.

Currently (commit ee7f4c047a4) the sidebar always shows it as active,
because the sidebar uses a weapon context that doesn't know which
hex it should be used in.