PROTOTYPE: Make normal weapon specials' terrain filters affect the sidebar
This starts as prototyping what would happen if just_teaching in
Newfrenchy's PR was implemented via is_for_listing_, then realising
that they should be the same thing, and then realising that normal
weapon filters should cause active/inactive colors to be used.

However, it's now broken the other way, because special_active() is
obeying is_for_listing_ and saying that the Dark Adept's special is
always active. That's a bug, that is_for_listing_ doesn't affect
the terrain filter.